Refactor code, move application logic from Tabs to Application
