+++ svelte/src/lib/components/Application.ts
@@ -3,14 +3,23 @@ import type { Action, IRmProjFile } from 'rockmelonqa.common';
 export const appActionContextKey = Symbol();
 
 export interface IAppActionContext {
-    getOnSaveHandler: (contentIndex: number) => OnSaveHandler | undefined;
+    /** Load selected *.rmproj file */
     loadProject: (projectFile: IRmProjFile) => void;
+
+    /** Open selected file in new tab */
     openFile: (relativeFilePath: string) => void;
+
+    // get, add, delete tab-content on-save handler
+    getOnSaveHandler: (contentIndex: number) => OnSaveHandler | undefined;
     registerOnSaveHandler: (contentIndex: number, func: OnSaveHandler) => void;
+    unregisterOnSaveHandler: (contentIndex: number) => void;
+        
     showCodeGenerationDialog: Action;
     showNewProjectDialog: Action;
     showRunTestDialog: Action;
-    unregisterOnSaveHandler: (contentIndex: number) => void;
+
+    /** Exit app */
+    quit: Action;
 }
 
 export type OnSaveHandler = () => Promise<boolean>;
